+++ src/astronomy/orbit.rs
@@ -1,4 +1,7 @@
+use std::time::Duration;
+
 use bevy::prelude::*;
+use bevy_spatial::{AutomaticUpdate, SpatialStructure};
 
 use crate::resources::{game_time::GameTime, state::GameState};
 
@@ -27,6 +30,12 @@ impl Default for Orbitable {
 pub struct OrbitPlugin;
 impl Plugin for OrbitPlugin {
     fn build(&self, app: &mut App) {
+        app.add_plugins(
+            // TODO: Replace use of Orbitable in KDTree with KDNode for reusability.
+            AutomaticUpdate::<Orbitable>::new()
+                .with_spatial_ds(SpatialStructure::KDTree2)
+                .with_frequency(Duration::from_millis(1)),
+        );
         app.add_systems(
             Update,
             (orbitable_update_system, orbital_positioning_system)